requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-dotenv
google-generativeai
